{merge,fmt-merge-msg}: Add the --from-name argument
We have a "--into-name" to fake one side of the merge message, but
there's no way to fake the from branch that's being merged.  The
specific case I'm running into is "b4 shazam -M" with shortlogs enabled,
which ends up putting a "/tmp/tmp.XXXXXX" as the from branch in the
shortlog (the title is already overriden by b4).

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>